Expose PSPThread in the same manner
